format

+++ Source/sidekick_state.h
@@ -12,5 +12,5 @@ struct SideKickState {
     Vec<T> vel_br;     // body relative axis rates
     Vec<T> accel_br;   // body relative axis accelerations
     int current_task;
-    unsigned long current_task_time;
+    uint32_t current_task_time;
 };
